Fix up downloadFileHns function and tests
