+++ index.test.js
@@ -5,7 +5,9 @@ describe("SkynetClient", () => {
     const client = new SkynetClient();
 
     // Download
+    expect(client).toHaveProperty("downloadData");
     expect(client).toHaveProperty("downloadFile");
+    expect(client).toHaveProperty("downloadFileHns");
     expect(client).toHaveProperty("getMetadata");
 
     // Encryption
@@ -16,6 +18,7 @@ describe("SkynetClient", () => {
     expect(client).toHaveProperty("getSkykeys");
 
     // Upload
+    expect(client).toHaveProperty("uploadData");
     expect(client).toHaveProperty("uploadFile");
     expect(client).toHaveProperty("uploadDirectory");
   });
